refactor: move main processing loop to new module
